Refactor AppEndProxy into domain-based partial classes
Split monolithic AppEndProxy.cs into multiple partial class files by domain (Ai, Scheduler, FileServices, DbServices, etc.) for better modularity and maintainability. Updated AppEndHost.csproj to include new files and remove obsolete ones. Clarified English-only code comment policy in copilot-instructions.md. No logic changes were made; this is an organizational refactor.
