+++ AppEndHost/AppEndHost.csproj
@@ -4,9 +4,9 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
     <RunPostBuildEvent>OnBuildSuccess</RunPostBuildEvent>
-    <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
-    <DefaultItemExcludes>$(DefaultItemExcludes);workspace\standalone\**\*;workspace\standalone\**\obj\**;workspace\standalone\**\bin\**</DefaultItemExcludes>
+    <SatelliteResourceLanguages>en</SatelliteResourceLanguages>    
   </PropertyGroup>
+
   <ItemGroup>
     <Content Remove="installedpackages.json" />
     <Content Remove="workspace\client\a..templates\BaseEmptyComponent.cshtml" />
@@ -23,6 +23,18 @@
     <Content Remove="workspace\client\a..templates\ReadTreeList.cshtml" />
     <Content Remove="workspace\client\a..templates\UpdateByKey.cshtml" />
     <Content Remove="workspace\client\Ali\app.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.ActivityLog.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.Ai.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.CacheServices.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.DbDialogServices.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.DbServices.DataSources.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.DbServices.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.DynaCodeServices.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.FileServices.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.HostingUtils.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.SchedulerApi.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.ServerActions.json" />
+    <Content Remove="workspace\server\Zzz.AppEndProxy.Uncategorized.json" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\AppEndCommon\AppEndCommon.csproj" />
@@ -79,6 +91,7 @@
     <Compile Remove="workspace\server\DefaultRepo.BaseUsers.cs" />
     <Compile Remove="workspace\server\DefaultRepo.BaseUsersRoles.cs" />
     <Compile Remove="workspace\server\Zzz.AppEndProxy.cs" />
+    <Compile Remove="workspace\server\Zzz.AppEndProxy.DbServices.DataSources.cs" />
     <None Remove="workspace\server\DefaultRepo.BaseActivityLog.cs" />
     <None Remove="workspace\server\DefaultRepo.BaseActivityLog.dbdialog.json" />
     <None Remove="workspace\server\DefaultRepo.BaseActivityLog.settings.json" />
@@ -115,7 +128,19 @@
     <None Remove="workspace\server\DefaultRepo.DbDirect.settings.json" />
     <None Remove="workspace\server\Zzz.Ai.cs" />
     <None Remove="workspace\server\Zzz.AppEndProxy.AAA.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.ActivityLog.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.Ai.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.CacheServices.cs" />
     <None Remove="workspace\server\Zzz.AppEndProxy.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.DbDialogServices.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.DbServices.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.DynaCodeServices.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.Exec.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.FileServices.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.HostingUtils.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.SchedulerApi.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.ServerActions.cs" />
+    <None Remove="workspace\server\Zzz.AppEndProxy.Settings.cs" />
     <None Remove="workspace\server\Zzz.AppEndProxy.settings.json" />
     <Content Include="workspace\server\Zzz.z.GlobalUsings.txt">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
@@ -156,7 +181,6 @@
     <Content Include="workspace\server\DefaultRepo.DbDirect.cs">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
-    <Content Include="workspace\server\Zzz.AppEndProxy.AAA.cs" />
     <Content Include="workspace\server\Zzz.AppEndProxy.cs">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
@@ -187,6 +211,45 @@
 
   <ItemGroup>
     <Compile Update="workspace\server\Zzz.Ai.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.AAA.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.ActivityLog.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.Ai.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.CacheServices.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.DbDialogServices.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.DbServices.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.DynaCodeServices.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.Exec.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.FileServices.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.HostingUtils.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.SchedulerApi.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.ServerActions.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Compile>
+    <Compile Update="workspace\server\Zzz.AppEndProxy.Settings.cs">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Compile>
   </ItemGroup>
