+++ AppEndHost/workspace/server/Zzz.AppEndProxy.ServerActions.cs
@@ -1,0 +1,17 @@
+namespace Zzz
+{
+    public static partial class AppEndProxy
+    {
+        #region ServerActions
+        public static object? RebuildProject()
+        {
+            return BuildServices.RebuildProject();
+        }
+        public static object? RestartApp()
+        {
+            
+            return true;
+        }
+        #endregion
+    }
+}
